Update MIDI.asm
Swapped around the logical tests for status byte such that the condition for if the status byte signals either a Note On or Off comes before the condition of if it is a Note Off. This does not affect the end functionality but gives the code a more logical flow.